PLATFORM-6134 | Apply alternative fix for T235188 cache corruption
getMultiWith[Union]SetCallback() defines a callback function to be invoked for
each cache key-ID pair in a loop, with the current ID value passed to the
function by reference and updated on each loop iteration. In the regular flow,
the key and ID will be correctly in sync; however, if a key is selected for
preemptive refresh, the callback will be stashed and kept around so that it can
be invoked again at the end of the request. Unfortunately, by this time the ID
reference will point to the ID of the last item in the key-ID map, since the
loop will have finished, and the preemptive refresh will execute with an
incorrect cache key-ID combo, resulting in cache corruption.

The upstream fix[1] is relatively convoluted and unfortunately does not apply
cleanly to 1.33. Let's fix the issue by simply creating a new callback for each
loop iteration and pass the current ID by value, rather than using a shared
callback and passing the ID by reference. While this theoretically has slightly
higher overhead (because we need to initialize a new closure on each iteration),
this added cost should be insignificant compared to the cost of memcached I/O
and regenerations on miss, and it helps keep the logic way simpler.

---
[1] https://gerrit.wikimedia.org/r/c/mediawiki/core/+/547696