PLATFORM-6677 | Fix revision_actor_temp locking on new revision creation
If multiple inserts happen at the same time, there are situations where upserting to new revision_actor_temp can cause deadlock. ActorMigration::getInsertValuesWithTempTable `pk` values are always autoincremented and there should not be cases where upsert is needed.

In one place where the function is used RevisionStore::insertRevisionRowOn, there is a similar function for comment on which insert is used as well.