PLATFORM-8338 | Add additional caching to the ChangeList rendering
Rendering UserLinks or Tags, requires parsing which when run on thousands of records, adds up to significant amount of processing time. More often then not, this can be optimized by storing already visited tags or user links as those "renders" will not change for the single request run but are repeated constantly on the list.