RecentChanges: Do not change "Create filter" button label from default
When a user saves current filter settings, the "apply" button's
message toggles whether the filter is marked as "default" or not.
As per the ticket, the message should always remain the same.

The width of the popup also changed to be wider to accommodate
a slightly longer title and the toggle of the "apply" button
width with the default icon.

Bug: T217304
Change-Id: Iee76cb422cbb011e89dd69d2feb0e74b30b2452c