MAIN-28662 | Vary RC user(talk) link cache on user name rather than ID
For anonymous users, the user ID is always zero, so it is possible for cached user links to be reused for different anonymous users after baf5c0a. This is not what we want, so use the user name as the cache key instead, which should be unique for both anonymous and registered users.