Disable automatic minification for UserOptionsModule
This module has highly dynamic, per-user content, which causes it to
generate an excessive amount of APCu cache entries if allowed to be
implicitly minified via the default ResourceLoader::filter() call in
makeModuleResponse(). Turn this off via FILTER_NOMIN and perform the
minification ourselves with caching disabled to avoid that.