Fix error in case where Main Page is set to User Page.
In case when wiki Main Page is set to any page from User: namespace
code will throw an error since it does not handle case when page from
user namespace can also be a Main Page. This fix will keep current logic
for associated pages and for user pages but will allow user page to be
set as a Main Page.